requirements.txt

Flask==3.0.0
Flask-Session==0.6.0
Werkzeug==3.0.1
python-dotenv==1.0.1

psycopg2-binary==2.9.9

gunicorn
